requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-Migrate==4.0.5
langchain==0.1.12
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
youtube-transcript-api==0.6.2
pytest==8.0.2
black==24.2.0
flake8==7.0.0
openai==1.12.0
google-api-python-client==2.118.0
google-search-results==2.4.2 